requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.1
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
